requirements.txt
flask
openai
cerebras_cloud_sdk
python-dotenv
fal-client
requests